NamedGraphs

Documentation for NamedGraphs.

NamedGraphs.DefaultNamedCapacity

NamedGraphs.Key

NamedGraphs.NamedDijkstraState

NamedGraphs.add_edges!

NamedGraphs.hexagonal_lattice_graph

NamedGraphs.incident_edges

NamedGraphs.is_path_graph

NamedGraphs.random_bfs_tree

NamedGraphs.rem_edges!

NamedGraphs.rem_vertices!

NamedGraphs.triangular_lattice_graph

NamedGraphs.DefaultNamedCapacity — Type

DefaultNamedCapacity{T}

Structure that returns 1 if a forward edge exists in flow_graph, and 0 otherwise.

NamedGraphs.Key — Type

Key{K}

A key (index) type, used for unambiguously identifying an object as a key or index of an indexible object AbstractArray, AbstractDict, etc.

Useful for nested structures of indices, for example:

[Key([1, 2]), [Key([3, 4]), Key([5, 6])]]

which could represent partitioning a set of vertices

[Key([1, 2]), Key([3, 4]), Key([5, 6])]

NamedGraphs.NamedDijkstraState — Type

struct NamedDijkstraState{V,T}

An AbstractPathState designed for Dijkstra shortest-paths calculations.

NamedGraphs.add_edges! — Method

Add a list of edges to a graph g

NamedGraphs.hexagonal_lattice_graph — Method

Generate a graph which corresponds to a hexagonal tiling of the plane. There are m rows and n columns of hexagons. Based off of the generator in Networkx hexagonallatticegraph()

NamedGraphs.incident_edges — Method

incident_edges(graph::AbstractGraph, vertex; dir=:out)

Edges incident to the vertex vertex.

dir ∈ (:in, :out, :both), defaults to :out.

For undirected graphs, returns all incident edges.

Like: https://juliagraphs.org/Graphs.jl/v1.7/algorithms/linalg/#Graphs.LinAlg.adjacency_matrix

NamedGraphs.is_path_graph — Method

TODO: Make this more sophisticated, check that only two vertices have degree 1 and none have degree 0, meaning it is a path/linear graph:

https://en.wikipedia.org/wiki/Path_graph

but not a path/linear forest:

https://en.wikipedia.org/wiki/Linear_forest

NamedGraphs.random_bfs_tree — Method

Do a BFS search to construct a tree, but do it with randomness to avoid generating the same tree. Based on Int. J. Comput. Their Appl. 15 pp 177-186 (2008). Edges will point away from source vertex s.

NamedGraphs.rem_edges! — Method

Remove a list of edges from a graph g

NamedGraphs.rem_vertices! — Method

Remove a list of vertices from a graph g

NamedGraphs.triangular_lattice_graph — Method

Generate a graph which corresponds to a equilateral triangle tiling of the plane. There are m rows and n columns of triangles. Based off of the generator in Networkx triangularlatticegraph()