Reference

BlockSparseArrays.BlockPermutedDiagonalAlgorithm — Type

BlockPermutedDiagonalAlgorithm(A::MatrixAlgebraKit.AbstractAlgorithm)

A wrapper for MatrixAlgebraKit.AbstractAlgorithm that implements the wrapped algorithm on a block-by-block basis, which is possible if the input matrix is a block-diagonal matrix or a block permuted block-diagonal matrix.

BlockSparseArrays.BlockPermutedDiagonalTruncationStrategy — Type

BlockPermutedDiagonalTruncationStrategy(strategy::TruncationStrategy)

A wrapper for TruncationStrategy that implements the wrapped strategy on a block-by-block basis, which is possible if the input matrix is a block-diagonal matrix or a block permuted block-diagonal matrix.

BlockSparseArrays.BlockSparseArray — Type

BlockSparseArray{T}(undef, dims)
BlockSparseArray{T,N}(undef, dims)
BlockSparseArray{T,N,A}(undef, dims)

Construct an uninitialized N-dimensional BlockSparseArray containing elements of type T. dims should be a list of block lengths in each dimension or a list of blocked ranges representing the axes.

SparseArraysBase.SparseArrayDOK — Method

SparseArrayDOK{T}(undef_blocks, axes)
SparseArrayDOK{T,N}(undef_blocks, axes)

Construct the block structure of an undefined BlockSparseArray that will have blocked axes axes.

Note that undef_blocks is defined in BlockArrays.jl and should be imported from that package to use it as an input to this constructor.

BlockSparseArrays.sparsemortar — Method

sparsemortar(blocks::AbstractArray{<:AbstractArray{T,N},N}, axes) -> ::BlockSparseArray{T,N}

Construct a block sparse array from a sparse array of arrays and specified blocked axes. The block sizes must be commensurate with the blocks of the axes.